fix travis test

+++ spec/relaton/db_cache_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe Relaton::DbCache do
     Relaton::DbCache.init_bib_caches(global_cache: true, local_cache: "", flush_caches: true)
     expect(File.exist?(File.expand_path("~/.relaton"))).to be true
     expect(File.exist?("relaton")).to be true
-    FileUtils.mv "relaton1", File.expand_path("~/.relaton")
+    FileUtils.mv "relaton1", File.expand_path("~/.relaton") if File.exist? "relaton1"
   end
 
   # it "returns fetched" do
